add minimal `fetch` wrapper for outbound HTTP, etc.
The main thing here is to add sdk.js, which currently only contains a `fetch`
function that provides a small subset of the standard JS `fetch` API, pretending
to be asynchronous but really just using the synchronous Spin outbound HTTP
`send` function.  Later, I'll add other wrappers to this file such as Node-style
filesystem functions.

This also reverts the QuickJS AOT-to-bytecode bits since profiling showed it
didn't make much difference speed-wise and only a modest difference memory- and
disk-wise.  Doesn't seem worth the extra complexity and build time.

Note that spin-js-engine had be be updated to handle the fact that
`handleRequest` now returns a `Promise`.  In general, this means we need to
flush the QuickJS job queue before we can be sure the `Promise` is resolved.
And then the only way to get the result is via a callback which stores it in a
global variable which we can finally return to the host.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>