fix error exit from `__wasilibc_populate_preopens`
For some reason, holding a `DirEntry` open indefinitely for a pre-opened
directory causes `__wasilibc_populate_preopens` to exit with an `EX_OSERR`
status.  That means we can't pass closures back to QuickJS which close over such
a `DirEntry`.  Instead, we get the `FileType` from the `DirEntry` and close over
that.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>