share the fermyon:spin@2.0.0 wit because otherwise wit merge will fail
This is sadly necessary because if each js package only defines its own
interface as its wit, current tooling for wit will refuse to merge them.

We can minimize the number of files in a future commit if required by
turning it into single file format but I am afraid that the duplication
is currently necessary.

Signed-off-by: karthik2804 <karthik.ganeshram@fermyon.com>