Fix sourcemap chaining
This commit fixes updating the sourcemap comment to point to the right
file. It also updates the test to use esbuild instead of webpack as the
sourcemaps generated by esbuild are friendlier without webpack speicif
protocol representing files. As we move to esbuild, the existing webpack
plugin is modified to suit esbuild as well.

Signed-off-by: Karthik Ganeshram <karthik.ganeshram@fermyon.com>